commit all to test on another machine
