+++ mkdocs.yml
@@ -15,11 +15,13 @@ nav:
     # - Avoidable Admissions: documentation/avoidable_admissions_sap.md
     # - Winter Pressures Statistical Analysis Plan: documentation/winter_pressures_sap.md
     - Global Variables Statistical Analysis Plan: documentation/global_variables.md
-  - Data Models: 
+  - Data Models:
     - Admitted Patient Care: data_models/admitted_patient_care_data_model.md
     - Emergency Care: data_models/emergency_care_data_model.md
-  - How To Guides: 
-    - New to Python?: how_to_guides/new_to_python.md
+  - How To Guides:
+    - Step 1) New to Python?: how_to_guides/new_to_python.md
+    - Step 2) From Terminal to Notebook/Lab: how_to_guides/from_terminal_to_notebook.md
+    - Step 3) From Notebook/Lab to Extract Complete: how_to_guides/from_notebook_to_data_extract.md
   - Code: code/code.md
 
 markdown_extensions:
