+++ modules/avoidable_admissions/__init__.py
@@ -1,0 +1,3 @@
+from avoidable_admissions import data, features, models, utils, visualization
+
+__all__ = ["data", "features", "models", "utils", "visualization"]
