+++ setup.py
@@ -1,0 +1,10 @@
+from setuptools import find_packages, setup
+
+setup(
+    name="avoidable_admissions",
+    packages=find_packages(),
+    version="0.1.0",
+    description="HDRUK Data Science Collaboration on Avoidable Admissions in the NHS",
+    author="LTHTR-DST",
+    license="MIT",
+)
